+++ tests/crawlers/test_cases/fileditch.py
@@ -4,7 +4,7 @@ TEST_CASES = [
         "url": "https://fileditchfiles.me/file.php?f=/b71/FrmLzfLKUHBWDTQfqaTZ.mp4",
         "results": [
             {
-                "url": "re:https://donotsharethesetemplinksyouidiot.st/b71/FrmLzfLKUHBWDTQfqaTZ.mp4",
+                "url": "re:/b71/FrmLzfLKUHBWDTQfqaTZ.mp4?md5=",
                 "filename": "FrmLzfLKUHBWDTQfqaTZ.mp4",
                 "referer": "https://fileditchfiles.me/file.php?f=/b71/FrmLzfLKUHBWDTQfqaTZ.mp4",
                 "download_folder": "re:Loose Files (Fileditch)",
@@ -17,7 +17,7 @@ TEST_CASES = [
         "description": "multiple slashes on URL",
         "results": [
             {
-                "url": "re:https://donotsharethesetemplinksyouidiot.st/b70/jRuBeGZlRoBWPurUARg.mp4",
+                "url": "re:/b70/jRuBeGZlRoBWPurUARg.mp4?md5=",
                 "referer": "https://fileditchfiles.me/file.php?f=/b70/jRuBeGZlRoBWPurUARg.mp4",
                 "download_folder": "re:Loose Files (Fileditch)",
                 "uploaded_at": None,
@@ -28,7 +28,7 @@ TEST_CASES = [
         "url": "https://fileditchfiles.me/beta5/a292619a708980582542/%5B8.11%5D_valk1.mp4",
         "results": [
             {
-                "url": "re:https://donotsharethesetemplinksyouidiot.st/beta5/a292619a708980582542/%5B8.11%5D_valk1.mp4",
+                "url": "re:/beta5/a292619a708980582542/%5B8.11%5D_valk1.mp4?md5=",
                 "filename": "[8.11]_valk1.mp4",
                 "original_filename": "[8.11]_valk1.mp4",
                 "referer": "https://fileditchfiles.me/beta5/a292619a708980582542/%5B8.11%5D_valk1.mp4",
@@ -40,25 +40,10 @@ TEST_CASES = [
         "count": 1,
     },
     {
-        "url": "https://fileditchfiles.me/temp/d9dc8e3cc669f9d24b16/3840x2839_b545565c05a233584ef041f248a38426.jpg",
-        "results": [
-            {
-                "url": "re:https://tempfileditchyall.me/temp/d9dc8e3cc669f9d24b16/3840x2839_b545565c05a233584ef041f248a38426.jpg",
-                "filename": "3840x2839_b545565c05a233584ef041f248a38426.jpg",
-                "original_filename": "3840x2839_b545565c05a233584ef041f248a38426.jpg",
-                "referer": "https://fileditchfiles.me/temp/d9dc8e3cc669f9d24b16/3840x2839_b545565c05a233584ef041f248a38426.jpg",
-                "album_id": None,
-                "uploaded_at": None,
-                "download_folder": "re:Loose Files (Fileditch)",
-            }
-        ],
-        "count": 1,
-    },
-    {
         "url": "https://fileditchfiles.me/alpha7/adf9b7514d0d86d1cadd/Anai_Loves_-_Cheating_At_The_World_Cup_Match_1080p.mp4",
         "results": [
             {
-                "url": "re:https://1.donotsharethesetemplinksyouidiot.st/alpha7/adf9b7514d0d86d1cadd/Anai_Loves_-_Cheating_At_The_World_Cup_Match_1080p.mp4",
+                "url": "re:/alpha7/adf9b7514d0d86d1cadd/Anai_Loves_-_Cheating_At_The_World_Cup_Match_1080p.mp4?md5=",
                 "filename": "Anai_Loves_-_Cheating_At_The_World_Cup_Match_1080p.mp4",
                 "original_filename": "Anai_Loves_-_Cheating_At_The_World_Cup_Match_1080p.mp4",
                 "referer": "https://fileditchfiles.me/alpha7/adf9b7514d0d86d1cadd/Anai_Loves_-_Cheating_At_The_World_Cup_Match_1080p.mp4",