+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - No videos detected on profiles, playlists or search results (Spankbang)
+- Download URL extraction (Fileditch)
 
 ## [10.0.2] - 2026-07-05